+++ src/main/java/stairwaytoheaven/biomes/StormveilBiome.java
@@ -10,11 +10,19 @@ import necesse.level.maps.biomes.MobSpawnTable;
 public class StormveilBiome extends SkyBiome {
 
     public static final MobSpawnTable mobs = new MobSpawnTable()
-            .add(70, "stormwisp")
-            .add(30, "zephyrray");
+            .addLimited(70, "stormwisp", 4, 80)
+            .addLimited(30, "zephyrray", 2, 80);
+
+    public static final MobSpawnTable critters = new MobSpawnTable()
+            .addLimited(100, "sparkbeetle", 4, 60);
 
     @Override
     public MobSpawnTable getMobSpawnTable(Level level) {
         return mobs;
+    }
+
+    @Override
+    public MobSpawnTable getCritterSpawnTable(Level level) {
+        return critters;
     }
 }